Work-in-progress version of networking foundations, and a few fixes in librmx

+++ librmx/source/rmxmedia/Font.cpp
@@ -328,7 +328,7 @@ void Font::printBitmap(Bitmap& outBitmap, Recti& outInnerRect, const StringReade
 	// Write inner rect
 	outInnerRect.x = -boundsMin.x;
 	outInnerRect.y = -boundsMin.y;
-	outInnerRect.width = (text.mLength - 1) * spacing;
+	outInnerRect.width = (int)(text.mLength - 1) * spacing;
 	for (size_t i = 0; i < text.mLength; ++i)
 	{
 		const FontSource::GlyphInfo* info = mFontSource->getGlyph(text[i]);
